requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.4
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
